Separate the two different views into individual packages.
